+++ src/lib.rs
@@ -9,14 +9,13 @@
 //! | COPY           | ✓             | ✓           |
 //! | LZMA           | ✓             | ✓           |
 //! | LZMA2          | ✓             | ✓           |
-//! | BROTLI (*/**)  | ✓             | ✓           |
+//! | BROTLI (*)     | ✓             | ✓           |
 //! | BZIP2 (*)      | ✓             | ✓           |
 //! | DEFLATE (*)    | ✓             | ✓           |
 //! | LZ4 (*)        | ✓             | ✓           |
 //! | ZSTD (*)       | ✓             | ✓           |
 //!
-//! (*) Require optional feature.
-//! (**) Currently we are not compatible with 7zip ZS when using the Brotli codec.
+//! (*) Require optional cargo feature.
 //!
 //! | Filter        | Decompression | Compression |
 //! |---------------|---------------|-------------|
@@ -35,6 +34,8 @@ extern crate wasm_bindgen;
 mod aes256sha256;
 mod bcj;
 mod bcj2;
+#[cfg(feature = "brotli")]
+mod brotli;
 #[cfg(not(target_arch = "wasm32"))]
 mod de_funcs;
 mod delta;
@@ -67,5 +68,6 @@ pub use reader::{BlockDecoder, SevenZReader};
 #[cfg(feature = "compress")]
 pub use writer::*;
 pub(crate) mod archive;
+
 pub(crate) mod decoders;
 pub(crate) mod folder;